Move code in `util` package to `_internal`
There was just a single function `cancel_and_await` that was only for
internal use.
