--- src/frequenz/sdk/util/__init__.py
@@ -1,4 +1,0 @@
-# License: MIT
-# Copyright © 2022 Frequenz Energy-as-a-Service GmbH
-
-"""General purpose tools."""
